Time: 0 ms (100.00%) | Memory: 8.5 MB (82.17%) - LeetSync

+++ solutions/3394-minimum-array-end/minimum-array-end.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    long long minEnd(int n, int x) {
+    long minEnd(int n, int x) {
         long result = x;
         long remaining = n - 1;
         long position = 1;
